Extract Endpoints object to its own file
